add float al posto di int per accettare anche i numeri decimali

+++ week1-python-basics/Giorno 5/esercizio_2_ciclo_for.py
@@ -1,10 +1,10 @@
 # esercizio 2: Numero più grande
 
-n = int(input("Quanti numeri inserirai? "))
+n = float(input("Quanti numeri inserirai? "))
 
-massimo = int(input("Scegli il numero: "))
+massimo = float(input("Scegli il numero: "))
 for i in range(n - 1):
-    x = int(input("Scegli un altro numero: "))
+    x = float(input("Scegli un altro numero: "))
     if massimo < x:
         massimo = x
 print(f"Il numero massimo che hai scelto è: {massimo}")
